Add Model.js file
- to separate music models
