+++ src/Scenes/TitleScene.js
@@ -58,6 +58,13 @@ export default class TitleScene extends Phaser.Scene {
     this.input.on('pointerout', function (event, gameObjects) {
       gameObjects[0].setTexture('blueButton1');
     });
+
+    this.model = this.sys.game.globals.model;
+    if (this.model.musicOn === true) {
+      this.bgMusic = this.sound.add('bgMusic', { volume: 0.5, loop: true });
+      this.bgMusic.play();
+    }
+
   }
 
   centerButton (gameObject, offset = 0) {
