+++ src/index.js
@@ -7,11 +7,17 @@ import PreloaderScene from './Scenes/PreloaderScene';
 import TitleScene from './Scenes/TitleScene';
 import OptionsScene from './Scenes/OptionsScene';
 import CreditsScene from './Scenes/CreditsScene';
-// import { SimpleScene } from './Scenes/simple-scene';
+import Model from './Model';
 
 class Game extends Phaser.Game {
   constructor(){
     super(config);
+
+    // call Model class
+    const model = new Model();
+    this.globals = { model };
+
+    // Call Scenes
     this.scene.add('Boot', BootScene);
     this.scene.add('Preloader', PreloaderScene);
     this.scene.add('Title', TitleScene);
